Producer/Consumer work in win64

+++ Launcher.h
@@ -12,7 +12,7 @@
 #endif
 
 class Launcher
-{	int pid;
+{	intptr_t pid;
 public:
 	Launcher()
 	:	pid(0)
